Update smoke with option bitcoin-reorg

+++ test/smoke/chains/bitcoin.py
@@ -94,7 +94,7 @@ class MockBitcoin:
         Get BTC balance for an address
         """
         unspents = self.call("listunspent", 1, 9999, [str(address)])
-        return int(sum(float(u["amount"]) for u in unspents) * Coin.ONE)
+        return int(sum(Decimal(u["amount"]) for u in unspents) * Coin.ONE)
 
     @retry(stop=stop_after_delay(30), wait=wait_fixed(1))
     def wait_for_node(self):
